Put the new daily process in place. + Move MainTest to MainIntegrationTest (and add real unit tests for Main in it's place) + Rename ModelRunManager & ModelRunGatekeeper to DiseaseProcess* + Move ModelRunManagerTest to ModelRunGatekeeperTest (now disease process) as that is what was really being tested. (unit tests for ModelRunManager/DiseaseProcessManager in next commit)
